+++ couplet_composer/support/tool_data.py
@@ -81,7 +81,8 @@ def list_tool_types():
         "make",
         "doxygen",
         "linter",
-        "linter_replacements"
+        "linter_replacements",
+        "coverage"
     ]
 
 
@@ -473,3 +474,18 @@ def create_clang_apply_replacements_tool_data(linter_required, tool_path=None):
                 ),
                 install_tool=lambda install_info, dry_run, print_debug: None
             )
+
+
+def create_lcov_tool_data():
+    """Creates the ToolData object of lcov for toolchain."""
+    return ToolData(
+        get_tool_key=lambda: "lcov",
+        get_tool_name=lambda: "lcov",
+        get_searched_tool=lambda: "lcov",
+        use_predefined_path=lambda: False,
+        get_required_local_version=lambda target, host_system: None,
+        get_local_executable=(
+                lambda tools_root, version, target, host_system: None
+            ),
+        install_tool=lambda install_info, dry_run, print_debug: None
+    )